requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coverage==4.5.2
coveralls==1.5.1
cryptography==2.4.2
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-cors-headers==2.4.0
django-extensions==2.1.4
django-rest-auth==0.9.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
flake8==3.6.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==5.0.0
oauthlib==2.1.0
pep8==1.7.1
pluggy==0.8.0
powerline-status===2.7.dev9999-git.b-5d82d544f3366a214732aeb8f781d50ced38ceef-
py==1.7.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
pytest==4.1.0
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
urllib3==1.24.1
wrapt==1.10.11
